GLOBAL: Start of input unification
