+++ Makefile.psp
@@ -30,7 +30,7 @@ COMMON_OBJS = \
 	source/platform/psp/battery.o \
 	source/platform/psp/thread.o \
 	source/platform/psp/VramExt.o \
-	source/platform/psp/input.o \
+	input.o \
 	source/platform/psp/main.o \
 	source/platform/psp/math.o \
 	source/platform/psp/sound.o \
@@ -140,7 +140,8 @@ LIBS		= $(GPROF_LIBS) $(GU_LIBS) $(AUDIO_LIBS) $(MISC_LIBS) $(STD_LIBS) $(NET_LI
 
 CFLAGS = -ffast-math -O3 -Ofast -G0 -Wall $(GPROF_FLAGS) -DPSP $(MODE) $(HARDWARE_VIDEO_ONLY_FLAGS) \
 		-DSWIZZLE32 -DPSP_MP3_HWDECODE -DFULLBRIGHT -DHL_RENDER -Wno-strict-aliasing -DPSP_VFPU -DPLATFORM_DIRECTORY="psp" \
-		-DPLATFORM_RENDERER="gu" -DMAX_AI_COUNT=12 -DRENDERER_SUPPORTS_DITHERING -DPLATFORM_USES_OSK
+		-DPLATFORM_RENDERER="gu" -DMAX_AI_COUNT=12 -DRENDERER_SUPPORTS_DITHERING -DPLATFORM_KEYBOARD_OSK \
+		-DPLATFORM_HAS_ONE_ANALOG_STICK
 
 ifeq ($(WERROR),1)
 CFLAGS 		+= -Werror
