+++ Makefile.psp2
@@ -100,7 +100,7 @@ COMMON_OBJS =	\
 	source/platform/psp2/net_psp2.o \
 	source/platform/psp2/net_adhoc_psp2.o \
 	source/platform/psp2/net_udp_psp2.o \
-	source/platform/psp2/in_psp2.o \
+	input.o \
 	source/platform/psp2/vgl/vgl_vidpsp2.o \
 	source/platform/psp2/neon_mathfun.o \
 	source/crypter.o
@@ -121,7 +121,7 @@ CFLAGS  = -fsigned-char -Wl,-q -O3 -g -fno-optimize-sibling-calls \
 	-DGLQUAKE -DHAVE_OGGVORBIS -DHAVE_MPG123 -DHAVE_LIBSPEEXDSP \
 	-DUSE_AUDIO_RESAMPLER -DHAVE_STRLCAT -DGIT_VERSION=\"$(GIT_VERSION)\" \
 	-DPLATFORM_DIRECTORY="psp2" -DPLATFORM_RENDERER="vgl" -DMAX_AI_COUNT=24 \
-	-DENABLE_RAZOR_CAPTURE -D__PSP2__ -DPLATFORM_USES_OSK
+	-DENABLE_RAZOR_CAPTURE -D__PSP2__ -DPLATFORM_KEYBOARD_OSK
 
 ifeq ($(WERROR),1)
 CFLAGS 		+= -Werror
