+++ source/platform/ctr/in_ctr.c
@@ -232,7 +232,7 @@ void IN_Move (usercmd_t *cmd)
 //
 // ctr software keyboard courtesy of libctru samples
 //
-void IN_SwitchKeyboard(void)
+void IN_OpenOSKeyboard(void)
 {
 	static SwkbdState swkbd;
 	static char console_buffer[64];
